Restructure as a standalone process.
This originated as part of my data QA script, but I'll ultimately need a
full-blown data cleaning pipeline - so let's get started.

clean_master.r loads up the data from csv and saves it back out to rdata,
running the numbered scripts in between.  This would be a little better as
a set of functions, but this works for now.