Use which() to fix indexing in age_check. Fixes #008
It's amazing that I've made it this far without discovering that trying to
index a data.frame using a logical comparison that contains NAs will totally
screw the data.frame.  I've been leaning on subset() quite heavily all these
years...

which() returns the numeric indices of the TRUE values in a logical vector,
but not the NAs - bypassing the problem pretty neatly.