Clarify WTF 'VisitDate' actually means on each table.
My naming convention here is that VisitDate should always mean the date of
whatever the table is recording - so "VisitDate" on the preenrollment table
means the preenrollment date, on the followupfortb table it means the follow-up,
etc. Thus, all the places where VisitDate has been tacked on by the DMS guys
are now "EnrollDate" - since they came off of the master table.